+++ Taller/Frac.h
@@ -1,6 +1,6 @@
 #ifndef FRAC_H_INCLUDED
 #define FRAC_H_INCLUDED
 
- double Frac( double x);
+ long double Frac( long double x);
 
 #endif // FRAC_H_INCLUDED
