+++ Taller/gmst.h
@@ -1,6 +1,6 @@
 #ifndef GMST_H_INCLUDED
 #define GMST_H_INCLUDED
 
- double gmst( double Mjd_UT1);
+ long double gmst( long double Mjd_UT1);
 
 #endif // GMST_H_INCLUDED
