+++ Taller/rpoly.h
@@ -1,6 +1,6 @@
 #ifndef RPOLY_H_INCLUDED
 #define RPOLY_H_INCLUDED
 
-int rpoly( double *op, int degree,  double *zeror,  double *zeroi, int info[]);
+int rpoly( long double *op, int degree,  long double *zeror,  long double *zeroi, int info[]);
 
 #endif
